refactor: Cambiar a aproximación preventiva en borrado de Food y Product
Anteriormente, se esperaba a que saltase DataIntegrityViolationException para ver si había dependencias ligadas a la entidad que se quería eliminar (aproximación reactiva).
Ahora se hace una comprobación preventiva, desde la lista de recetas de dicha entidad, para ver si tiene dependencias ligadas o no. En caso de que sí, ahí es cuando se lanza la excepción (no se depende de que salte la excepción de la BBDD ni se espera al rollback).

+++ app/spring-boot/src/main/java/es/ubu/inf/tfg/food/FoodService.java
@@ -2,7 +2,6 @@ package es.ubu.inf.tfg.food;
 
 import java.util.List;
 
-import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import es.ubu.inf.tfg.exception.ResourceInUseException;
@@ -80,11 +79,10 @@ public class FoodService { // TODO: cambiar con DTO?
     public void delete(Integer id) {
 
         Food food = findById(id);
-        try {
-            foodRepository.delete(food);
+
+        if (!food.getRecipes().isEmpty()) {
+            throw new ResourceInUseException("Food", id, "Product");
         }
-        catch (DataIntegrityViolationException e) {
-            throw new ResourceInUseException("Food", id);
-        }
+        foodRepository.delete(food);
     }
 }
